+++ packages/rosaenlg-no-pug/package.json
@@ -6,7 +6,7 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "clean": "exit 0 && rm -rf dist",
-    "test": "exit 0 && nyc --reporter=lcov --reporter=text mocha",
+    "test": "nyc --reporter=lcov --reporter=text mocha",
     "build": "exit 0 && tsc"
   },
   "dependencies": {
